refactor(sync): split sync communication service into modular components
break down the monolithic sync communication service into focused
helper classes for better maintainability and separation of concerns

- SyncDtoSerializer handles DTO-to-JSON conversion with isolate support
- SyncMessageSerializer manages WebSocket message serialization
- SyncWebSocketClient handles connection management and communication
- Updated import path in sync registration module
